+++ app/(admin)/index.tsx
@@ -5,7 +5,7 @@ import {
   StyleSheet,
 } from 'react-native'
 
-import CardList from '@/components/ui/dashboard/admin/card-list'
+import CardList from '@/components/ui/screens/dashboard/admin/card-list'
 import TransactionItem from '@/components/ui/transaction/transaction-item'
 import SkeletonContent from '@alanantar_sb/react-native-skeleton-content'
 import { DateFormatter } from '@/utils/types'
